[ANI-39-radio-group] refactor: type을 분류합니다.
