+++ apps/web/app/page.tsx
@@ -1,7 +1,8 @@
 'use client'
 
-import { renderComponents, type ServerDrivenComponentType } from '@repo/ui/renderComponents'
+import { renderComponents } from '@repo/ui/renderComponents'
 import example from './example.json' assert { type: 'json' }
+import { type ServerDrivenComponentType } from '@repo/ui/types'
 
 export default function Home() {
   return <div>{renderComponents(example as ServerDrivenComponentType)}</div>
