+++ packages/ui/src/components/renderComponents.tsx
@@ -1,18 +1,5 @@
-import { ComponentProps } from 'react'
 import RadioGroup from './atom/radio-group/RadioGroup'
-
-type ComponentPropsMap = {
-  RadioGroup: ComponentProps<typeof RadioGroup>
-  RadioGroupItem: ComponentProps<typeof RadioGroup.Item>
-}
-
-export type ServerDrivenComponentType =
-  | {
-      type: 'RadioGroup'
-      props: ComponentPropsMap['RadioGroup']
-      children?: ServerDrivenComponentType[]
-    }
-  | { type: 'RadioGroupItem'; props: ComponentPropsMap['RadioGroupItem']; children?: never }
+import { ServerDrivenComponentType } from '@/types/server-driven'
 
 export function renderComponents(componentData: ServerDrivenComponentType) {
   const { type, props, children } = componentData
