+++ packages/ui/src/types/server-driven.ts
@@ -1,0 +1,15 @@
+import RadioGroup from '@/components/atom/radio-group/RadioGroup'
+import { ComponentProps } from 'react'
+
+export type ComponentPropsMap = {
+  RadioGroup: ComponentProps<typeof RadioGroup>
+  RadioGroupItem: ComponentProps<typeof RadioGroup.Item>
+}
+
+export type ServerDrivenComponentType =
+  | {
+      type: 'RadioGroup'
+      props: ComponentPropsMap['RadioGroup']
+      children?: ServerDrivenComponentType[]
+    }
+  | { type: 'RadioGroupItem'; props: ComponentPropsMap['RadioGroupItem']; children?: never }
